create non-linked thumbnail component
