+++ app/components/searchworks4/link_thumbnail_component.html.erb
